requirements.txt
pandas
pymongo
openpyxl
xlrd
python-dotenv
fastapi
uvicorn
numpy